+++ src/pages/petition-notice/edit/page.tsx
@@ -1,10 +1,11 @@
 import { Header } from '@/containers/common/Header/Header';
 import { PetitionNoticeEditorSection } from './containers/PetitionNoticeEditorSection';
+import { State } from '@/containers/common/Header/const/state';
 
 export function PetitionNoticeEditPage() {
   return (
     <div>
-      <Header />
+      <Header state={State.Login} />
       <PetitionNoticeEditorSection />
     </div>
   );
