+++ src/style.css
@@ -155,12 +155,14 @@ body {
 }
 
 .toastui-editor-contents p {
-  font-weight: 500;
+  font-family: 'Pretendard';
+  font-weight: 400;
   color: #7e7e7e;
 }
 
 .toastui-editor-contents h3 {
-  font-weight: 900;
+  font-family: 'Pretendard';
+  font-weight: 700;
   color: #484848;
 }
 
@@ -172,7 +174,7 @@ body {
   font-size: 19px;
 }
 
-@media screen and (min-width: 720px) {
+@media screen and (min-width: 390px) {
   .toastui-editor-contents p {
     font-size: 18px;
   }
